requirements.txt
Package            Version
------------------ -------
click              8.1.5
Flask              2.2.5
Flask-Cors         4.0.0
Flask-JWT-Extended 4.5.2
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.3
pip                23.1.2
psycopg2-binary    2.9.6
PyJWT              2.7.0
python-dotenv      1.0.0
setuptools         68.0.0
Werkzeug           2.3.6
wheel              0.40.0
